Module cayley::path

Structs

Morphism
An interface to a Path with the ability to be saved and reused to construct other Paths, but not to query anything.

Vertex
An interface to a Path with the ability to be executed as a query to a database. The main entry point to ask for GraphNodes from database using Graph as an interceptor.

Traits

Compile
Marks any Path which is able to be compiled to a string Gremlin-compatible query

Path
The trait which covers all the methods from Gremlin API Path specification, but in a Rust way.

Query

Reuse