Struct cayley::path::Morphism

pub struct Morphism {
    // some fields omitted
}

An interface to a Path with the ability to be saved and reused to construct other Paths, but not to query anything.

Use it to prepare a Path and re-use it several times

#![allow(unused_result)]
use cayley::{Graph, DefaultVersion};
use cayley::path::Vertex as V;
use cayley::path::Morphism as M;
use cayley::path::{Path, Query}; // both traits imports are required
use cayley::selector::{Predicate, Node};

let graph = Graph::new("localhost", 64210, DefaultVersion).unwrap();
let mut follows_m = M::start("foo");
        follows_m.OutP(Predicate("follows"));
graph.save(&mut follows_m).unwrap();
graph.find(V::start(Node("C"))
             .Follow(&follows_m)
             .Has(Predicate("status"), Node("cool_person"))
             .All()).unwrap();

Methods

impl Morphism

fn start(name: &str) -> Morphism

Create a Morphism instance with intention to store it in database under the given name